JENKINS-246 Parallelize the stages to decrease build times to 15 minutes.
At the moment only two stages will run in parallel.
Their work load is split to roughly equal, that means one is dominated by
the PR test suite while the other stage executes the rest.

Unfortunately I was not able to use a top-level agent and a different agent
in one of the parallel stages.
As a result, there is no top-level agent.
Without a top-level agent the last post-directive need dedicated nodes
to execute on. I choose master since the steps executed are quite cheap performance-wise.

Unfortunately it was also not possible to share the agent-part due to the
restrictive checks employed by the declarative pipeline syntax at the time
of this commit. Thus the agent-block is duplicated multiple times.